chore(app/settings): rename settings tabs components
