+++ src/components/navigation/navbar.tsx
@@ -45,10 +45,7 @@ const Navbar = () => {
 
         <Dialog open={isOpen} onOpenChange={setIsOpen}>
           <DialogTrigger asChild>
-            <Button
-              variant="ghost"
-              className="no-focus cursor-pointer hover:bg-transparent! [&_svg:not([class*='size-'])]:size-10!"
-            >
+            <Button variant="ghost" className="no-focus hover:bg-transparent">
               <MenuIcon className="size-12" />
             </Button>
           </DialogTrigger>
