+++ src/modules/app/settings/tabs/standard-settings.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { Plus } from 'lucide-react';
 import { PlayerAvatar } from '../../hud/player-avatar';
 import { avatarGroup, playerGroup } from '../constant/form';
-import { CommonSettings } from './common';
+import { CommonSettings } from './common-settings';
 import { useSettings } from '@/contexts/settings';
 import { useAppForm } from '@/components/form/hooks';
 import { StandardSettingsSchema } from '@/lib/validations';
